+++ packages/zowe-explorer/__tests__/__unit__/utils/UriHandler.unit.test.ts
@@ -11,25 +11,37 @@
 
 import { commands, Uri } from "vscode";
 import { ZoweUriHandler } from "../../../src/utils/UriHandler";
+import { DatasetFSProvider } from "../../../src/trees/dataset/DatasetFSProvider";
 
 describe("ZoweUriHandler", () => {
     function getBlockMocks() {
         return {
-            executeCommand: jest.spyOn(commands, "executeCommand"),
+            remoteLookupForResource: jest.spyOn(DatasetFSProvider.instance, "remoteLookupForResource"),
         };
     }
 
     it("does nothing if the parsed query does not start with a Zowe scheme", async () => {
         const blockMocks = getBlockMocks();
         await ZoweUriHandler.getInstance().handleUri(Uri.parse("vscode://Zowe.vscode-extension-for-zowe?blah-some-unknown-query"));
-        expect(blockMocks.executeCommand).not.toHaveBeenCalled();
+        expect(blockMocks.remoteLookupForResource).not.toHaveBeenCalled();
     });
 
-    it("calls vscode.open with the parsed URI if a Zowe resource URI was provided", async () => {
+    it("calls remoteLookupForResource with the parsed URI if a Zowe resource URI was provided", async () => {
         const blockMocks = getBlockMocks();
         const uri = Uri.parse("vscode://Zowe.vscode-extension-for-zowe?zowe-ds:/lpar.zosmf/TEST.PS");
         await ZoweUriHandler.getInstance().handleUri(uri);
         const zoweUri = Uri.parse(uri.query);
-        expect(blockMocks.executeCommand).toHaveBeenCalledWith("vscode.open", zoweUri, { preview: false });
+        expect(blockMocks.remoteLookupForResource).toHaveBeenCalledWith(zoweUri);
+    });
+
+    it("calls remoteLookupForResource with the parsed URI if a Zowe resource URI was provided", async () => {
+        const blockMocks = getBlockMocks();
+        blockMocks.remoteLookupForResource.mockResolvedValue({ name: "exampleEntry" } as any);
+        const executeCommandSpy = jest.spyOn(commands, "executeCommand");
+        const uri = Uri.parse("vscode://Zowe.vscode-extension-for-zowe?zowe-ds:/lpar.zosmf/TEST.PS");
+        await ZoweUriHandler.getInstance().handleUri(uri);
+        const zoweUri = Uri.parse(uri.query);
+        expect(blockMocks.remoteLookupForResource).toHaveBeenCalledWith(zoweUri);
+        expect(executeCommandSpy).toHaveBeenCalledWith("vscode.open", zoweUri, { preview: false });
     });
 });